requirements.txt
appnope==0.1.0
backcall==0.1.0
bleach==3.0.0
certifi==2018.8.24
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
elasticsearch==6.3.1
entrypoints==0.2.3
etk==2.1.5
Flask==1.0.2
idna==2.7
ipykernel==5.0.0
ipython==7.0.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.1
Jinja2==2.10
jsonlines==1.2.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
langdetect==1.0.7
MarkupSafe==1.0
mistune==0.8.3
nbconvert==5.4.0
nbformat==4.4.0
nltk==3.4
notebook==5.7.2
numpy==1.15.2
pandas==0.22.0
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prometheus-client==0.4.0
prompt-toolkit==2.0.5
psycopg2==2.7.5
psycopg2-binary==2.7.5
ptyprocess==0.6.0
Pygments==2.2.0
pyparsing==2.3.0
python-dateutil==2.7.3
pytz==2018.5
pyzmq==17.1.2
qtconsole==4.4.1
regex<2017.12.1,>=2017.4.5
requests==2.20.0
scikit-learn==0.20.1
selenium==3.141.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
spacy==2.0.17
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
tld==0.9.1
tornado==5.1.1
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
Wikipedia-API==0.3.7
wget==3.2
widgetsnbextension==3.4.2
xlrd==1.1.0
